+++ src/main/kotlin/io/github/yuitosato/kotlintree/dsl/TreeNodeDSL.kt
@@ -1,4 +1,9 @@
-package io.github.yuitosato.kotlintree
+package io.github.yuitosato.kotlintree.dsl
+
+import io.github.yuitosato.kotlintree.MutableTreeNode
+import io.github.yuitosato.kotlintree.TreeNode
+import io.github.yuitosato.kotlintree.asMutable
+import io.github.yuitosato.kotlintree.leafOf
 
 /**
  * Returns a tree node in DSL style.
